+++ src/test/java/redis/clients/jedis/tests/commands/AllKindOfValuesCommandsTest.java
@@ -37,7 +37,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
         status = jedis.set(bfoo, bbar);
         assertEquals("OK", status);
 
-        int reply = jedis.exists("foo");
+        long reply = jedis.exists("foo");
         assertEquals(1, reply);
 
         reply = jedis.exists(bfoo);
@@ -62,7 +62,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
         jedis.set("foo2", "bar2");
         jedis.set("foo3", "bar3");
 
-        int reply = jedis.del("foo1", "foo2", "foo3");
+        long reply = jedis.del("foo1", "foo2", "foo3");
         assertEquals(3, reply);
 
         reply = jedis.exists("foo1");
@@ -222,7 +222,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
     @Test
     public void renamenx() {
         jedis.set("foo", "bar");
-        int status = jedis.renamenx("foo", "bar");
+        long status = jedis.renamenx("foo", "bar");
         assertEquals(1, status);
 
         jedis.set("foo", "bar");
@@ -231,7 +231,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
 
         // Binary
         jedis.set(bfoo, bbar);
-        int bstatus = jedis.renamenx(bfoo, bbar);
+        long bstatus = jedis.renamenx(bfoo, bbar);
         assertEquals(1, bstatus);
 
         jedis.set(bfoo, bbar);
@@ -242,7 +242,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
 
     @Test
     public void dbSize() {
-        int size = jedis.dbSize();
+        long size = jedis.dbSize();
         assertEquals(0, size);
 
         jedis.set("foo", "bar");
@@ -257,7 +257,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
 
     @Test
     public void expire() {
-        int status = jedis.expire("foo", 20);
+        long status = jedis.expire("foo", 20);
         assertEquals(0, status);
 
         jedis.set("foo", "bar");
@@ -265,7 +265,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
         assertEquals(1, status);
 
         // Binary
-        int bstatus = jedis.expire(bfoo, 20);
+        long bstatus = jedis.expire(bfoo, 20);
         assertEquals(0, bstatus);
 
         jedis.set(bfoo, bbar);
@@ -278,7 +278,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
     public void expireAt() {
         long unixTime = (System.currentTimeMillis() / 1000L) + 20;
 
-        int status = jedis.expireAt("foo", unixTime);
+        long status = jedis.expireAt("foo", unixTime);
         assertEquals(0, status);
 
         jedis.set("foo", "bar");
@@ -287,7 +287,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
         assertEquals(1, status);
 
         // Binary
-        int bstatus = jedis.expireAt(bfoo, unixTime);
+        long bstatus = jedis.expireAt(bfoo, unixTime);
         assertEquals(0, bstatus);
 
         jedis.set(bfoo, bbar);
@@ -299,7 +299,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
 
     @Test
     public void ttl() {
-        int ttl = jedis.ttl("foo");
+        long ttl = jedis.ttl("foo");
         assertEquals(-1, ttl);
 
         jedis.set("foo", "bar");
@@ -311,7 +311,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
         assertTrue(ttl >= 0 && ttl <= 20);
 
         // Binary
-        int bttl = jedis.ttl(bfoo);
+        long bttl = jedis.ttl(bfoo);
         assertEquals(-1, bttl);
 
         jedis.set(bfoo, bbar);
@@ -345,7 +345,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
 
     @Test
     public void move() {
-        int status = jedis.move("foo", 1);
+        long status = jedis.move("foo", 1);
         assertEquals(0, status);
 
         jedis.set("foo", "bar");
@@ -358,7 +358,7 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
 
         // Binary
         jedis.select(0);
-        int bstatus = jedis.move(bfoo, 1);
+        long bstatus = jedis.move(bfoo, 1);
         assertEquals(0, bstatus);
 
         jedis.set(bfoo, bbar);
@@ -428,14 +428,14 @@ public class AllKindOfValuesCommandsTest extends JedisCommandTestBase {
     public void persist() {
         jedis.setex("foo", 60 * 60, "bar");
         assertTrue(jedis.ttl("foo") > 0);
-        int status = jedis.persist("foo");
+        long status = jedis.persist("foo");
         assertEquals(1, status);
         assertEquals(-1, jedis.ttl("foo").intValue());
 
         // Binary
         jedis.setex(bfoo, 60 * 60, bbar);
         assertTrue(jedis.ttl(bfoo) > 0);
-        int bstatus = jedis.persist(bfoo);
+        long bstatus = jedis.persist(bfoo);
         assertEquals(1, bstatus);
         assertEquals(-1, jedis.ttl(bfoo).intValue());
