+++ src/test/java/redis/clients/jedis/tests/commands/BinaryValuesCommandsTest.java
@@ -75,7 +75,7 @@ public class BinaryValuesCommandsTest extends JedisCommandTestBase {
 
     @Test
     public void setnx() {
-        int status = jedis.setnx(bfoo, binaryValue);
+        long status = jedis.setnx(bfoo, binaryValue);
         assertEquals(1, status);
         assertTrue(Arrays.equals(binaryValue, jedis.get(bfoo)));
 
@@ -88,7 +88,7 @@ public class BinaryValuesCommandsTest extends JedisCommandTestBase {
     public void setex() {
         String status = jedis.setex(bfoo, 20, binaryValue);
         assertEquals(Keyword.OK.name(), status);
-        int ttl = jedis.ttl(bfoo);
+        long ttl = jedis.ttl(bfoo);
         assertTrue(ttl > 0 && ttl <= 20);
     }
 
@@ -102,7 +102,7 @@ public class BinaryValuesCommandsTest extends JedisCommandTestBase {
 
     @Test
     public void msetnx() {
-        int status = jedis.msetnx(bfoo, binaryValue, bbar, bfoo);
+        long status = jedis.msetnx(bfoo, binaryValue, bbar, bfoo);
         assertEquals(1, status);
         assertTrue(Arrays.equals(binaryValue, jedis.get(bfoo)));
         assertTrue(Arrays.equals(bfoo, jedis.get(bbar)));
@@ -121,7 +121,7 @@ public class BinaryValuesCommandsTest extends JedisCommandTestBase {
 
     @Test
     public void incr() {
-        int value = jedis.incr(bfoo);
+        long value = jedis.incr(bfoo);
         assertEquals(1, value);
         value = jedis.incr(bfoo);
         assertEquals(2, value);
@@ -135,7 +135,7 @@ public class BinaryValuesCommandsTest extends JedisCommandTestBase {
 
     @Test
     public void incrBy() {
-        int value = jedis.incrBy(bfoo, 2);
+        long value = jedis.incrBy(bfoo, 2);
         assertEquals(2, value);
         value = jedis.incrBy(bfoo, 2);
         assertEquals(4, value);
@@ -149,7 +149,7 @@ public class BinaryValuesCommandsTest extends JedisCommandTestBase {
 
     @Test
     public void decr() {
-        int value = jedis.decr(bfoo);
+        long value = jedis.decr(bfoo);
         assertEquals(-1, value);
         value = jedis.decr(bfoo);
         assertEquals(-2, value);
@@ -163,7 +163,7 @@ public class BinaryValuesCommandsTest extends JedisCommandTestBase {
 
     @Test
     public void decrBy() {
-        int value = jedis.decrBy(bfoo, 2);
+        long value = jedis.decrBy(bfoo, 2);
         assertEquals(-2, value);
         value = jedis.decrBy(bfoo, 2);
         assertEquals(-4, value);
@@ -173,7 +173,7 @@ public class BinaryValuesCommandsTest extends JedisCommandTestBase {
     public void append() {
         byte[] first512 = new byte[512];
         System.arraycopy(binaryValue, 0, first512, 0, 512);
-        int value = jedis.append(bfoo, first512);
+        long value = jedis.append(bfoo, first512);
         assertEquals(512, value);
         assertTrue(Arrays.equals(first512, jedis.get(bfoo)));
